widget_api: rename driver -> matrix mod
