+++ crates/matrix-sdk/src/widget_api/mod.rs
@@ -2,9 +2,13 @@
 #![allow(missing_docs)]
 
 pub mod api;
-pub mod driver;
 pub mod error;
 pub mod handler;
+pub mod matrix;
 pub mod messages;
 
 pub use self::error::{Error, Result};
+
+pub fn run_client_widget_api() {
+    println!("Hello, world!");
+}
